Fix templates so they work with new StrictUndefined setting for rendering
In some cases, we want template rendering to fail if a variable is not defined.  In other cases, we want to check for definition, and if it's not defined, proceed without executing a certain part of the template.  Specifying the latter as {% if key is defined %} rather than {% if key %}  allows us to do that.